# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jengo Building Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List residents
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create a resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "201":
          description: Created resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResidentBody"
      responses:
        "200":
          description: Updated resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"

  # ── Issues ─────────────────────────────────────────────────────────────────
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: category
          in: query
          schema:
            type: string
            enum: [noise, maintenance, parking, visitor, utility, security, other]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create an issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueBody"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update issue status/resolution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueBody"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/comments:
    get:
      operationId: listIssueComments
      tags: [issues]
      summary: List comments on an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IssueComment"
    post:
      operationId: addIssueComment
      tags: [issues]
      summary: Add a comment to an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueComment"
  /issues/summary:
    get:
      operationId: getIssuesSummary
      tags: [issues]
      summary: Get issues summary (counts by status and category)
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Issues summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuesSummary"

  # ── Announcements ──────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List service charge payments
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue, waived]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/summary:
    get:
      operationId: getPaymentsSummary
      tags: [payments]
      summary: Get payments summary (collected, overdue, outstanding by building)
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsSummary"

  # ── Contractors ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Resident:
      type: object
      properties:
        id:
          type: integer
        unitId:
          type: integer
        buildingId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        moveInDate:
          type: string
          format: date
        isOwner:
          type: boolean
        status:
          type: string
          enum: [active, inactive, moved_out]
        createdAt:
          type: string
          format: date-time
      required: [id, unitId, buildingId, firstName, lastName, phone, status, createdAt]
    CreateResidentBody:
      type: object
      properties:
        unitId:
          type: integer
        buildingId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        moveInDate:
          type: string
          format: date
        isOwner:
          type: boolean
      required: [unitId, buildingId, firstName, lastName, phone]
    UpdateResidentBody:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive, moved_out]

    # ── Issue ──
    Issue:
      type: object
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [noise, maintenance, parking, visitor, utility, security, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        evidenceUrl:
          type: string
        evidenceType:
          type: string
          enum: [photo, audio, video, document]
        assignedTo:
          type: string
        resolutionNote:
          type: string
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, buildingId, title, category, priority, status, createdAt]
    CreateIssueBody:
      type: object
      properties:
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [noise, maintenance, parking, visitor, utility, security, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        evidenceUrl:
          type: string
        evidenceType:
          type: string
          enum: [photo, audio, video, document]
      required: [buildingId, title, category, priority]
    UpdateIssueBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        assignedTo:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        resolutionNote:
          type: string
    IssueComment:
      type: object
      properties:
        id:
          type: integer
        issueId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
          enum: [resident, management, caretaker, security]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, issueId, authorName, authorRole, content, createdAt]
    CreateIssueCommentBody:
      type: object
      properties:
        authorName:
          type: string
        authorRole:
          type: string
          enum: [resident, management, caretaker, security]
        content:
          type: string
      required: [authorName, authorRole, content]
    IssuesSummary:
      type: object
      properties:
        totalOpen:
          type: integer
        totalInProgress:
          type: integer
        totalResolved:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        byPriority:
          type: array
          items:
            type: object
            properties:
              priority:
                type: string
              count:
                type: integer
            required: [priority, count]
        avgResolutionHours:
          type: number
      required: [totalOpen, totalInProgress, totalResolved, byCategory, byPriority]

    # ── Announcement ──
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: KES
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue, waived]
        paymentMethod:
          type: string
          enum: [mpesa, bank_transfer, cash, other]
        mpesaRef:
          type: string
        month:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, buildingId, unitId, description, amount, dueDate, status, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        month:
          type: string
      required: [buildingId, unitId, description, amount, dueDate]
    UpdatePaymentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, overdue, waived]
        paidDate:
          type: string
          format: date
        paymentMethod:
          type: string
          enum: [mpesa, bank_transfer, cash, other]
        mpesaRef:
          type: string
    PaymentsSummary:
      type: object
      properties:
        totalCollected:
          type: number
        totalOutstanding:
          type: number
        totalOverdue:
          type: number
        collectionRate:
          type: number
        overdueCount:
          type: integer
        paidCount:
          type: integer
        pendingCount:
          type: integer
      required: [totalCollected, totalOutstanding, totalOverdue, collectionRate, overdueCount, paidCount, pendingCount]

    # ── Contractor ──
...